Move createSession to sessions/core/
