+++ packages/core/src/sessions/core/create-session.spec.ts
@@ -1,9 +1,12 @@
+// std
 import { notStrictEqual, ok, strictEqual } from 'assert';
-import { createService } from '../core';
+
+// FoalTS
+import { createService } from '../../core';
 import { createSession } from './create-session';
-import { Session } from './session';
-import { SessionState } from './session-state.interface';
-import { SessionStore } from './session-store';
+import { Session } from '../session';
+import { SessionState } from '../session-state.interface';
+import { SessionStore } from '../session-store';
 
 describe('createSession', () => {
