+++ packages/core/src/sessions/use-sessions.hook.ts
@@ -18,7 +18,7 @@ import {
   ServiceManager
 } from '../core';
 import { SESSION_DEFAULT_COOKIE_NAME } from './constants';
-import { createSession } from './create-session';
+import { createSession } from './core/create-session';
 import { checkUserIdType } from './http/check-user-id-type';
 import { getSessionIDFromRequest, RequestValidationError } from './http/get-session-id-from-request';
 import { readSession } from './read-session';
